swift: split audio and image decoding out of Inputs.swift
