+++ bindings/swift/Sources/RunAnywhere/Public/API/Inputs.swift
@@ -152,97 +152,6 @@ public struct AudioInput: Sendable {
         return source
     }
 
-    /// Mono 16-bit PCM plus the rate it is at.
-    ///
-    /// Reading and decoding audio is platform I/O, which is the SDK's job under
-    /// the layering rules; commons has neither a decoder nor a file reader.
-    private static func decodeToPCM16(_ input: AudioInput) throws -> (samples: Data, sampleRate: Int) {
-        #if canImport(AVFoundation)
-        let url: URL
-        var temporary: URL?
-        defer { if let temporary { try? FileManager.default.removeItem(at: temporary) } }
-
-        switch input.encoding {
-        case .file:
-            guard let path = input.path else {
-                throw SDKException(code: .invalidInput, message: "AudioInput.file has no path",
-                                   category: .validation)
-            }
-            url = URL(fileURLWithPath: path)
-        default:
-            // AVAudioFile reads a file, so buffer-borne container bytes get one.
-            let scratch = FileManager.default.temporaryDirectory
-                .appendingPathComponent(UUID().uuidString)
-                .appendingPathExtension("wav")
-            try input.data.write(to: scratch)
-            temporary = scratch
-            url = scratch
-        }
-
-        let file: AVAudioFile
-        do {
-            file = try AVAudioFile(forReading: url)
-        } catch {
-            throw SDKException(code: .invalidInput,
-                               message: "could not read audio at \(url.lastPathComponent): \(error)",
-                               category: .validation)
-        }
-
-        let rate = file.processingFormat.sampleRate
-        guard let target = AVAudioFormat(commonFormat: .pcmFormatInt16, sampleRate: rate,
-                                         channels: 1, interleaved: true),
-              let converter = AVAudioConverter(from: file.processingFormat, to: target) else {
-            throw SDKException(code: .invalidInput, message: "unsupported audio format",
-                               category: .validation)
-        }
-
-        var samples = Data()
-        let chunkFrames: AVAudioFrameCount = 16384
-        while file.framePosition < file.length {
-            guard let input = AVAudioPCMBuffer(pcmFormat: file.processingFormat,
-                                               frameCapacity: chunkFrames) else { break }
-            try file.read(into: input)
-            if input.frameLength == 0 { break }
-
-            let capacity = AVAudioFrameCount(
-                (Double(input.frameLength) * rate / file.processingFormat.sampleRate).rounded(.up) + 1
-            )
-            guard let output = AVAudioPCMBuffer(pcmFormat: target, frameCapacity: capacity) else { break }
-            var pending: NSError?
-            var supplied = false
-            converter.convert(to: output, error: &pending) { _, status in
-                if supplied {
-                    status.pointee = .noDataNow
-                    return nil
-                }
-                supplied = true
-                status.pointee = .haveData
-                return input
-            }
-            if let pending {
-                throw SDKException(code: .invalidInput,
-                                   message: "could not decode audio: \(pending.localizedDescription)",
-                                   category: .validation)
-            }
-            if let channel = output.int16ChannelData?[0], output.frameLength > 0 {
-                samples.append(UnsafeBufferPointer(start: channel, count: Int(output.frameLength)))
-            }
-        }
-
-        guard !samples.isEmpty else {
-            throw SDKException(code: .invalidInput, message: "audio decoded to no samples",
-                               category: .validation)
-        }
-        return (samples, Int(rate))
-        #else
-        throw SDKException(
-            code: .featureNotAvailable,
-            message: "Decoding audio files needs AVFoundation; supply AudioInput.pcm16 instead",
-            category: .validation
-        )
-        #endif
-    }
-
 
     func toVADAudioSource() throws -> RAVADAudioSource {
         var source = RAVADAudioSource()
@@ -516,78 +425,6 @@ public struct ImageInput: Sendable {
     }
 
     // MARK: Decoding
-
-    /// RAVLMImageFormat was deleted along with the rest of the closed VLM
-    /// image-format enum (idl/vlm_options.proto); `mediaType` is now a plain
-    /// MIME string, so magic-byte sniffing resolves directly to one instead
-    /// of to an enum case.
-    private static func mediaType(of data: Data) -> String {
-        guard data.count >= 4 else { return "application/octet-stream" }
-        let prefix = [UInt8](data.prefix(4))
-        if prefix[0] == 0xFF, prefix[1] == 0xD8 { return "image/jpeg" }
-        if prefix[0] == 0x89, prefix[1] == 0x50 { return "image/png" }
-        if prefix[0] == 0x52, prefix[1] == 0x49 { return "image/webp" }
-        return "application/octet-stream"
-    }
-
-    private static func decode(_ data: Data) throws -> (data: Data, width: Int, height: Int) {
-        #if canImport(CoreGraphics)
-        guard let source = CGImageSourceCreateWithData(data as CFData, nil),
-              let image = CGImageSourceCreateImageAtIndex(source, 0, nil),
-              let rgb = packedRGB(from: image) else {
-            throw SDKException(
-                code: .invalidInput,
-                message: "Could not decode the supplied image bytes",
-                category: .validation
-            )
-        }
-        return (rgb, image.width, image.height)
-        #else
-        throw SDKException(
-            code: .featureNotAvailable,
-            message: "Image decoding needs CoreGraphics; supply ImageInput.rawRgb instead",
-            category: .validation
-        )
-        #endif
-    }
-
-    #if canImport(CoreGraphics)
-    private static func packedRGB(from image: CGImage) -> Data? {
-        let width = image.width
-        let height = image.height
-        let bytesPerRow = 4 * width
-        let totalBytes = bytesPerRow * height
-        guard totalBytes > 0 else { return nil }
-
-        var rgba = Data(count: totalBytes)
-        var drew = false
-        rgba.withUnsafeMutableBytes { buffer in
-            guard let context = CGContext(
-                data: buffer.baseAddress,
-                width: width,
-                height: height,
-                bitsPerComponent: 8,
-                bytesPerRow: bytesPerRow,
-                space: CGColorSpaceCreateDeviceRGB(),
-                bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue
-            ) else { return }
-            context.draw(image, in: CGRect(x: 0, y: 0, width: width, height: height))
-            drew = true
-        }
-        guard drew else { return nil }
-
-        var rgb = Data(capacity: width * height * 3)
-        rgba.withUnsafeBytes { buffer in
-            let pixels = buffer.bindMemory(to: UInt8.self)
-            for index in stride(from: 0, to: totalBytes, by: 4) {
-                rgb.append(pixels[index])
-                rgb.append(pixels[index + 1])
-                rgb.append(pixels[index + 2])
-            }
-        }
-        return rgb
-    }
-    #endif
 
     /// Encode packed 24-bit RGB pixels as a PNG container, for diffusion's
     /// `image`/`mask_image` fields (which require an encoded container, not
